Moved multiple files from /graphics to /scene level
