+++ src/deprecated/deprecated.js
@@ -39,10 +39,10 @@ import { drawQuadWithShader } from '../graphics/simple-post-effect.js';
 import { shaderChunks } from '../scene/shader-lib/chunks/chunks.js';
 import { GraphicsDevice } from '../graphics/graphics-device.js';
 import { IndexBuffer } from '../graphics/index-buffer.js';
-import { createFullscreenQuad, drawFullscreenQuad, PostEffect } from '../graphics/post-effect.js';
+import { createFullscreenQuad, drawFullscreenQuad, PostEffect } from '../scene/graphics/post-effect.js';
 import { PostEffectQueue } from '../framework/components/camera/post-effect-queue.js';
-import { ProgramLibrary } from '../graphics/program-library.js';
-import { getProgramLibrary, setProgramLibrary } from '../graphics/get-program-library.js';
+import { ProgramLibrary } from '../scene/shader-lib/program-library.js';
+import { getProgramLibrary, setProgramLibrary } from '../scene/shader-lib/get-program-library.js';
 import { RenderTarget } from '../graphics/render-target.js';
 import { ScopeId } from '../graphics/scope-id.js';
 import { Shader } from '../graphics/shader.js';
