+++ src/index.js
@@ -58,16 +58,16 @@ export { Ray } from './core/shape/ray.js';
 // GRAPHICS
 export * from './graphics/constants.js';
 export { drawQuadWithShader, drawTexture } from './graphics/simple-post-effect.js';
-export { shFromCubemap } from './graphics/prefilter-cubemap.js';
-export { reprojectTexture } from './graphics/reproject-texture.js';
+export { shFromCubemap } from './scene/graphics/prefilter-cubemap.js';
+export { reprojectTexture } from './scene/graphics/reproject-texture.js';
 export { createShader, createShaderFromCode } from './scene/shader-lib/utils.js';
 export { shaderChunks } from './scene/shader-lib/chunks/chunks.js';
 export { shaderChunksLightmapper } from './scene/shader-lib/chunks/chunks-lightmapper.js';
 export { GraphicsDevice } from './graphics/graphics-device.js';
-export { EnvLighting } from './graphics/env-lighting.js';
+export { EnvLighting } from './scene/graphics/env-lighting.js';
 export { IndexBuffer } from './graphics/index-buffer.js';
-export { PostEffect, drawFullscreenQuad } from './graphics/post-effect.js';
-export { ProgramLibrary } from './graphics/program-library.js';
+export { PostEffect, drawFullscreenQuad } from './scene/graphics/post-effect.js';
+export { ProgramLibrary } from './scene/shader-lib/program-library.js';
 export { RenderTarget } from './graphics/render-target.js';
 export { ScopeId } from './graphics/scope-id.js';
 export { ScopeSpace } from './graphics/scope-space.js';
