+++ src/scene/graphics/env-lighting.js
@@ -1,13 +1,13 @@
-import { Vec4 } from '../core/math/vec4.js';
-import { Texture } from './texture.js';
+import { Vec4 } from '../../core/math/vec4.js';
+import { Texture } from '../../graphics/texture.js';
 import { reprojectTexture } from './reproject-texture.js';
 import {
     TEXTURETYPE_DEFAULT, TEXTURETYPE_RGBM,
     TEXTUREPROJECTION_EQUIRECT,
     ADDRESS_CLAMP_TO_EDGE,
     PIXELFORMAT_R8_G8_B8_A8, PIXELFORMAT_RGBA16F, PIXELFORMAT_RGBA32F
-} from './constants.js';
-import { DebugGraphics } from './debug-graphics.js';
+} from '../../graphics/constants.js';
+import { DebugGraphics } from '../../graphics/debug-graphics.js';
 
 const fixCubemapSeams = true;
 const RGBA8_TYPE = TEXTURETYPE_RGBM;
