+++ src/scene/graphics/prefilter-cubemap.js
@@ -1,14 +1,14 @@
-import { Debug } from '../core/debug.js';
-import { Vec3 } from '../core/math/vec3.js';
+import { Debug } from '../../core/debug.js';
+import { Vec3 } from '../../core/math/vec3.js';
 
 import {
     PIXELFORMAT_R8_G8_B8_A8, TEXTURETYPE_DEFAULT, TEXTURETYPE_RGBM
-} from './constants.js';
-import { createShaderFromCode } from '../scene/shader-lib/utils.js';
-import { drawQuadWithShader } from './simple-post-effect.js';
-import { shaderChunks } from '../scene/shader-lib/chunks/chunks.js';
-import { RenderTarget } from './render-target.js';
-import { Texture } from './texture.js';
+} from '../../graphics/constants.js';
+import { createShaderFromCode } from '../shader-lib/utils.js';
+import { drawQuadWithShader } from '../../graphics/simple-post-effect.js';
+import { shaderChunks } from '../shader-lib/chunks/chunks.js';
+import { RenderTarget } from '../../graphics/render-target.js';
+import { Texture } from '../../graphics/texture.js';
 
 // https://seblagarde.wordpress.com/2012/06/10/amd-cubemapgen-for-physically-based-rendering/
 function areaElement(x, y) {
