+++ src/scene/graphics/reproject-texture.js
@@ -1,22 +1,22 @@
-import { Debug } from '../core/debug.js';
+import { Debug } from '../../core/debug.js';
 import {
     FILTER_NEAREST,
     TEXTUREPROJECTION_OCTAHEDRAL, TEXTUREPROJECTION_CUBE
-} from './constants.js';
-import { Vec3 } from '../core/math/vec3.js';
-import { random } from '../core/math/random.js';
-import { createShaderFromCode } from '../scene/shader-lib/utils.js';
-import { drawQuadWithShader } from './simple-post-effect.js';
-import { ChunkUtils } from '../scene/shader-lib/chunk-utils.js';
-import { shaderChunks } from '../scene/shader-lib/chunks/chunks.js';
-import { RenderTarget } from './render-target.js';
-import { GraphicsDevice } from './graphics-device.js';
-import { getProgramLibrary } from './get-program-library.js';
-import { Texture } from './texture.js';
-import { DebugGraphics } from './debug-graphics.js';
-import { DeviceCache } from './device-cache.js';
+} from '../../graphics/constants.js';
+import { Vec3 } from '../../core/math/vec3.js';
+import { random } from '../../core/math/random.js';
+import { createShaderFromCode } from '../shader-lib/utils.js';
+import { drawQuadWithShader } from '../../graphics/simple-post-effect.js';
+import { ChunkUtils } from '../shader-lib/chunk-utils.js';
+import { shaderChunks } from '../shader-lib/chunks/chunks.js';
+import { RenderTarget } from '../../graphics/render-target.js';
+import { GraphicsDevice } from '../../graphics/graphics-device.js';
+import { getProgramLibrary } from '../shader-lib/get-program-library.js';
+import { Texture } from '../../graphics/texture.js';
+import { DebugGraphics } from '../../graphics/debug-graphics.js';
+import { DeviceCache } from '../../graphics/device-cache.js';
 
-/** @typedef {import('../core/math/vec4.js').Vec4} Vec4 */
+/** @typedef {import('../../core/math/vec4.js').Vec4} Vec4 */
 
 const getProjectionName = (projection) => {
     switch (projection) {
