+++ src/scene/materials/basic-material.js
@@ -7,7 +7,7 @@ import {
 
 import { basic } from '../../scene/shader-lib/programs/basic.js';
 import { ShaderProcessorOptions } from '../../graphics/shader-processor-options.js';
-import { getProgramLibrary } from '../../graphics/get-program-library.js';
+import { getProgramLibrary } from '../shader-lib/get-program-library.js';
 import { Material } from './material.js';
 
 /** @typedef {import('../../graphics/texture.js').Texture} Texture */
